test-requirements.txt
# Copyright 2021 Canonical Ltd.
# See LICENSE file for licensing details.

flake8==4.0.1
black
flake8-copyright<0.3
juju==3.0.1
pytest
pyyaml
lightkube==0.11.0